refactor(#4712): pure-move de-giant discord/meeting_orchestrator.rs
